Use asynchronous calls
The original implementation used the synchronous KuduClient, which
would block the server every time a call was made. To prevent this,
we should use AsyncKuduClient, which requires the use of callbacks.
Right now, these callbacks do very little, but in future work, I
would like them to:

1. try again on failure (maybe have background thread running queued-
   up failed tasks, limit 3 per customer)
2. return a response on failure (making all requests GET requests)
3. do nothing on failure (pretty much what happens now)

Note: I've made scan synchronous. Not sure if it makes sense
	  otherwise, unless we send a PUT request back to the client,
	  but that would put more work on users.